Rename nosqlstore > mapstore
