+++ mapstore-core/src/main/java/net/spals/appbuilder/mapstore/core/model/MapRangeKey.java
@@ -1,0 +1,11 @@
+package net.spals.appbuilder.mapstore.core.model;
+
+/**
+ * @author tkral
+ */
+public interface MapRangeKey<C extends Comparable<C>> {
+
+   MapRangeOperator getOperator();
+
+   C getValue();
+}
