--- nosqlstore-core/src/main/java/net/spals/appbuilder/nosqlstore/core/model/NoSqlStoreRangeKey.java
@@ -1,11 +1,0 @@
-package net.spals.appbuilder.nosqlstore.core.model;
-
-/**
- * @author tkral
- */
-public interface NoSqlStoreRangeKey<C extends Comparable<C>> {
-
-   NoSqlStoreOperator getOperator();
-
-   C getValue();
-}
